requirements.txt
transformers
bitsandbytes
accelerate
peft
datasets
scipy
tyro
python-dotenv
jsonformer
streamlit
selenium
Pillow